+++ test/e2e/suites.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package framework
+package e2e
 
 import (
 	"fmt"
@@ -22,7 +22,7 @@ import (
 	"path"
 	"time"
 
-	// TODO: Remove the following imports (ref: https://github.com/kubernetes/kubernetes/issues/81245)
+	"k8s.io/kubernetes/test/e2e/framework"
 	e2emetrics "k8s.io/kubernetes/test/e2e/framework/metrics"
 )
 
@@ -32,36 +32,36 @@ import (
 // and then the function that only runs on the first Ginkgo node.
 func CleanupSuite() {
 	// Run on all Ginkgo nodes
-	Logf("Running AfterSuite actions on all nodes")
-	RunCleanupActions()
+	framework.Logf("Running AfterSuite actions on all nodes")
+	framework.RunCleanupActions()
 }
 
 // AfterSuiteActions are actions that are run on ginkgo's SynchronizedAfterSuite
 func AfterSuiteActions() {
 	// Run only Ginkgo on node 1
-	Logf("Running AfterSuite actions on node 1")
-	if TestContext.ReportDir != "" {
-		CoreDump(TestContext.ReportDir)
+	framework.Logf("Running AfterSuite actions on node 1")
+	if framework.TestContext.ReportDir != "" {
+		framework.CoreDump(framework.TestContext.ReportDir)
 	}
-	if TestContext.GatherSuiteMetricsAfterTest {
+	if framework.TestContext.GatherSuiteMetricsAfterTest {
 		if err := gatherTestSuiteMetrics(); err != nil {
-			Logf("Error gathering metrics: %v", err)
+			framework.Logf("Error gathering metrics: %v", err)
 		}
 	}
-	if TestContext.NodeKiller.Enabled {
-		close(TestContext.NodeKiller.NodeKillerStopCh)
+	if framework.TestContext.NodeKiller.Enabled {
+		close(framework.TestContext.NodeKiller.NodeKillerStopCh)
 	}
 }
 
 func gatherTestSuiteMetrics() error {
-	Logf("Gathering metrics")
-	c, err := LoadClientset()
+	framework.Logf("Gathering metrics")
+	c, err := framework.LoadClientset()
 	if err != nil {
 		return fmt.Errorf("error loading client: %v", err)
 	}
 
 	// Grab metrics for apiserver, scheduler, controller-manager, kubelet (for non-kubemark case) and cluster autoscaler (optionally).
-	grabber, err := e2emetrics.NewMetricsGrabber(c, nil, !ProviderIs("kubemark"), true, true, true, TestContext.IncludeClusterAutoscalerMetrics)
+	grabber, err := e2emetrics.NewMetricsGrabber(c, nil, !framework.ProviderIs("kubemark"), true, true, true, framework.TestContext.IncludeClusterAutoscalerMetrics)
 	if err != nil {
 		return fmt.Errorf("failed to create MetricsGrabber: %v", err)
 	}
@@ -73,13 +73,13 @@ func gatherTestSuiteMetrics() error {
 
 	metricsForE2E := (*e2emetrics.ComponentCollection)(&received)
 	metricsJSON := metricsForE2E.PrintJSON()
-	if TestContext.ReportDir != "" {
-		filePath := path.Join(TestContext.ReportDir, "MetricsForE2ESuite_"+time.Now().Format(time.RFC3339)+".json")
+	if framework.TestContext.ReportDir != "" {
+		filePath := path.Join(framework.TestContext.ReportDir, "MetricsForE2ESuite_"+time.Now().Format(time.RFC3339)+".json")
 		if err := ioutil.WriteFile(filePath, []byte(metricsJSON), 0644); err != nil {
 			return fmt.Errorf("error writing to %q: %v", filePath, err)
 		}
 	} else {
-		Logf("\n\nTest Suite Metrics:\n%s\n", metricsJSON)
+		framework.Logf("\n\nTest Suite Metrics:\n%s\n", metricsJSON)
 	}
 
 	return nil
